+++ backend/server/internal/responses/user_details_response.go
@@ -9,7 +9,7 @@ type (
 		Username		string 		`json:"username"`
 		UserRank		string		`json:"user_rank"`
 		UserScore		uint64		`json:"user_score"`
-		UserGlobalRank uint64	`json:"user_global_score"`
+		UserGlobalRank float32 `json:"user_global_score"`
 		UserSolvedProblems uint64 `json:"user_solved_problems"`
 
 		SolvedProblemsList []ProblemBrief `json:"solved_problems_list"`
